Separate the neutral spec-input gates from the Fortran runner emitter.
`tools/runner_renderer.py` carried two things that are not language knowledge: the
two node-IDENTITY spec-input gates (`spec_id_length_violation` /
`infra_dep_count_violation`) and the case_id safe-token grammar. Both run where no
language has been resolved — the closure builder applies the gates to `spec_ref`s
out of `deps.yaml`, before an IR exists — so neither can be answered by asking the
registry which backend a node uses.

Move them to `tools/spec_input_gates.py`, a neutral module. The spec_id bound stays
a neutral number rather than a question put to a backend, for that reason; the next
commit pins it against the language backend's own identifier limit so the two cannot
drift. The case_id grammar becomes public, which retires three cross-module imports
of a private name (the conductor's argv builder, the pipeline validator's case-id
gate, the build-runtime server's test).

The gate messages lose their Fortran wording: they are read by an author of a spec
that has no language yet.
