+++ tools/runner_renderer.py
@@ -34,8 +34,12 @@ the *certified* harness IR signatures + source before rendering.
 
 from __future__ import annotations
 
-import re
 from typing import Any
+
+# Neutral spec-input policy this emitter also enforces at render time: the safe-token grammar a
+# case_id must obey (it reaches a filesystem path and an argv, neither of which is a language
+# question). One owner, three readers — see `tools/spec_input_gates.py`.
+from tools.spec_input_gates import CASE_ID_TOKEN_RE
 
 # The fixed ABI of the leaf-authored `<spec_id>_checks` module (see
 # docs/workflow/CHECKS_MODULE_CONTRACT.md). Non-prefixed public names (module
@@ -71,91 +75,11 @@ CHECK_STATUS_WIDTH = 4
 CASE_ID_LEN = 64
 
 # f2008 identifier limit; the longest derived name is `<spec_id>_checks` /
-# `<spec_id>_runner` (spec_id + 7). Keep a margin: cap spec_id at 55.
+# `<spec_id>_runner` (spec_id + 7). Keep a margin: cap spec_id at 55. The neutral spec-input
+# gate (`tools/spec_input_gates.py`) carries the same bound as a pre-IR precondition; this copy
+# is the render-time backstop.
 MAX_SPEC_ID_LEN = 55
 MAX_IDENTIFIER_LEN = 63
-
-# The character grammar a case_id must obey. The harness builds each per-case snapshot path by
-# concatenating the runtime case_id — `raw/state_snapshots/'//trim(case_id)//'.json'` (harness
-# controlled_spec §3) — so a `case_id` such as `../../evil` traverses OUT of the run directory
-# and the runner (which compiles and runs cleanly) writes an arbitrary file. The compile gates
-# only require a case_id to be a non-empty string, and `_flit` only bars non-printable-ASCII, so
-# `..`/`/` slip through both. Restrict the id to the same safe-token grammar the dependency layer
-# uses for path segments (`orchestration_runtime._is_safe_path_token`): `[A-Za-z0-9._-]`, no `..`,
-# narrowed further below because a case id also reaches the runner's argv.
-# The first character additionally may not be `-`: a case id reaches the runner's argv
-# through `--cases $(SPEC) $(CASES)` and the build-runtime MCP server refuses a leading
-# `-` there, so accepting one here would pass Compile and Build and then fail
-# Validate.execute on an id no gate had objected to.
-_CASE_ID_TOKEN_RE = re.compile(r"^[A-Za-z0-9._][A-Za-z0-9._-]*$")
-
-
-def spec_id_length_violation(spec_id: Any) -> str | None:
-    """Spec-input bound on spec_id length — the M3d mass-opt-in prerequisite gate.
-
-    Returns an actionable violation message when ``spec_id`` exceeds
-    ``MAX_SPEC_ID_LEN``, else ``None``. On a make+fortran node the derived
-    ``<spec_id>_runner``/``_checks``/``_model`` identifiers (spec_id + 7) breach the
-    f2008 ``MAX_IDENTIFIER_LEN``-char limit, and on a harness-backed M3c node the
-    host-rendered runner additionally fail-closes at ``_check_identifier_lengths``
-    (a workflow-kill a compile.generate re-author cannot repair — the spec_id is
-    node IDENTITY, not authored IR content). This helper is the canonical
-    *spec-input* capture point for exactly that identity precondition, which the
-    compile.static hoist deliberately excludes: bounding here — before any phase
-    runs — turns an unrepairable late render-kill into an early, clear rejection.
-    The renderer keeps the same bound as a defense-in-depth backstop."""
-    sid = spec_id.strip() if isinstance(spec_id, str) else ""
-    if len(sid) > MAX_SPEC_ID_LEN:
-        return (
-            f"spec_id {sid!r} is {len(sid)} chars (>{MAX_SPEC_ID_LEN}); the derived "
-            f"`<spec_id>_runner`/`_checks`/`_model` identifiers would breach the f2008 "
-            f"{MAX_IDENTIFIER_LEN}-char limit (and fail-close a harness-backed node's "
-            f"host-render). Rename the spec to ≤{MAX_SPEC_ID_LEN} chars.")
-    return None
-
-
-def infra_dep_count_violation(spec_kind: Any, infra_dep_count: int) -> str | None:
-    """Spec-input bound on the number of ``infrastructure`` direct dependencies.
-
-    Returns an actionable violation message unless the node declares EXACTLY ONE
-    ``infrastructure`` (runner-harness) direct dependency, or is itself an
-    ``infrastructure`` spec (the harness authors its own self-test runner, so it
-    declares none). Sibling of ``spec_id_length_violation``: both are node-IDENTITY
-    preconditions a Compile re-author cannot repair, so both are captured at
-    spec-input rather than hoisted into the compile.static gate (routing an
-    unrepairable defect to a warm-resume retry would only spin).
-
-    Before this gate, a physics node with zero or >1 infrastructure deps silently
-    degraded to the leaf-authored-runner path: ``_conductor_authors_runner`` requires
-    exactly one, so the runner was simply never host-rendered and the failure was a
-    quiet loss of the harness path rather than an error. That non-M3c physical path
-    has been removed — the only live leaf-authored runner is an ``infrastructure``
-    node's own self-test — so the degradation is now a hard rejection."""
-    # `.strip()` and NOTHING else — the exemption must be spelled exactly as every
-    # downstream reader spells it. `_conductor_authors_runner`, `_pure_leaf_substep` and
-    # `_validate_toolchain_backend_supported` all compare `str(...).strip() ==
-    # "infrastructure"` with no case folding, so a `spec_kind: Infrastructure` that this
-    # gate lower-cased into an exemption would be treated as a PHYSICS node by all three —
-    # exempted here and then silently landed on the removed leaf-authored-runner path, with
-    # no gate firing anywhere. Being case-sensitive here makes that shape a spec-input
-    # rejection instead, which is the direction that fails closed.
-    kind = spec_kind.strip() if isinstance(spec_kind, str) else ""
-    if kind == "infrastructure":
-        return None
-    if infra_dep_count == 1:
-        return None
-    remedy = (
-        "Add the single `infrastructure_id` entry" if infra_dep_count < 1
-        else f"Remove {infra_dep_count - 1} of them, keeping the one harness this node "
-             "builds against")
-    return (
-        f"a non-infrastructure spec must declare exactly one `infrastructure` "
-        f"(runner-harness) dependency in deps.yaml; found {infra_dep_count}. The runner "
-        f"glue is host-rendered against exactly that harness, and the former "
-        f"leaf-authored-runner path for a node without it has been removed "
-        f"(docs/workflow/phases/phase_01_compile.md). {remedy} "
-        f"(see spec/problem/dynamics/advection_diffusion/advdiff1d_linear/deps.yaml)."
-    )
 
 
 # The deterministic Generate.gate lint-checker column limit (fortitude S001). The rendered runner must stay
@@ -272,7 +196,7 @@ def _case_ids(ir: dict[str, Any]) -> list[str]:
     # traverses out of the run directory and the (cleanly compiling) runner writes an arbitrary
     # file. `_flit` only bars non-printable-ASCII and the length gate only bounds size, so both
     # let `../evil` through. Restrict the id to a filesystem-and-Fortran-safe token.
-    unsafe = sorted({c for c in out if not _CASE_ID_TOKEN_RE.match(c) or ".." in c})
+    unsafe = sorted({c for c in out if not CASE_ID_TOKEN_RE.match(c) or ".." in c})
     if unsafe:
         raise RenderError(
             f"case_id(s) {unsafe} are not safe tokens; a case_id is concatenated into the "
