+++ tools/tests/test_runner_renderer.py
@@ -24,14 +24,11 @@ from tools.runner_renderer import (
     CHECK_STATUS_WIDTH,
     EXPECTED_HARNESS_SPEC_ID,
     _HARNESS_V3_PARAMETERS,
-    MAX_SPEC_ID_LEN,
     RenderError,
     _HARNESS_V3_INTERFACE,
     assert_harness_pin,
-    infra_dep_count_violation,
     ir_content_violations,
     render_runner,
-    spec_id_length_violation,
 )
 from tools.backends.language.fortran.signatures import parse_signatures_from_fortran
 
@@ -1040,96 +1037,6 @@ class IrContentViolationsTest(unittest.TestCase):
             path(ir)
             v = ir_content_violations(ir, BOUNDARY_SID, HARNESS)
             self.assertTrue(v, v)  # non-empty, and no exception escaped
-
-
-class SpecIdLengthGateTest(unittest.TestCase):
-    """M3d spec-input gate: `spec_id_length_violation` bounds spec_id ≤ MAX_SPEC_ID_LEN.
-    This is the canonical capture point for the one node-IDENTITY render precondition the
-    compile.static hoist excludes (a re-author cannot shorten a spec_id); the conductor's
-    resolve_node calls it so a too-long spec_id fails before any phase, not at render-kill."""
-
-    def test_ok_at_or_below_limit(self) -> None:
-        self.assertIsNone(spec_id_length_violation("shallow_water2d"))
-        self.assertIsNone(spec_id_length_violation("x" * MAX_SPEC_ID_LEN))
-        # 54-char real catalog id (the longest currently-safe one) passes.
-        self.assertIsNone(
-            spec_id_length_violation("dynamics_advection_diffusion_boundary_1d_periodic_copy"))
-
-    def test_violation_above_limit(self) -> None:
-        # The 61-char id the catalog carried until it was renamed to fit the bound.
-        offender = "dynamics_advection_diffusion_profile_1d_upwind_center2_euler1"
-        msg = spec_id_length_violation(offender)
-        self.assertIsNotNone(msg)
-        self.assertIn(str(len(offender)), msg)
-        self.assertIn(str(MAX_SPEC_ID_LEN), msg)
-        # One char over the limit is already a violation; the limit itself is not.
-        self.assertIsNotNone(spec_id_length_violation("y" * (MAX_SPEC_ID_LEN + 1)))
-        self.assertIsNone(spec_id_length_violation("y" * MAX_SPEC_ID_LEN))
-
-    def test_non_string_and_whitespace(self) -> None:
-        self.assertIsNone(spec_id_length_violation(None))
-        self.assertIsNone(spec_id_length_violation(123))
-        # Surrounding whitespace is stripped before measuring.
-        self.assertIsNone(spec_id_length_violation("  short  "))
-        self.assertIsNotNone(spec_id_length_violation("  " + "z" * (MAX_SPEC_ID_LEN + 1) + "  "))
-
-
-class InfraDepCountGateTest(unittest.TestCase):
-    """Spec-input gate: `infra_dep_count_violation` requires EXACTLY ONE `infrastructure`
-    direct dependency on every non-infrastructure spec. Sibling of the spec_id bound: both
-    are node-IDENTITY preconditions a Compile re-author cannot repair, so both are captured
-    at spec-input. Zero and >1 used to degrade silently to the removed leaf-authored-runner
-    path; they are hard rejections now."""
-
-    def test_exactly_one_passes(self) -> None:
-        for kind in ("component", "profile", "problem"):
-            self.assertIsNone(infra_dep_count_violation(kind, 1), kind)
-
-    def test_zero_and_more_than_one_violate(self) -> None:
-        for kind in ("component", "profile", "problem"):
-            for count in (0, 2, 3):
-                self.assertIsNotNone(infra_dep_count_violation(kind, count), (kind, count))
-
-    def test_infrastructure_kind_is_exempt_at_every_count(self) -> None:
-        # The harness authors its own self-test runner, so it declares no harness of its own.
-        for count in (0, 1, 5):
-            self.assertIsNone(infra_dep_count_violation("infrastructure", count), count)
-        self.assertIsNone(infra_dep_count_violation("  infrastructure  ", 0))
-
-    def test_the_exemption_is_case_sensitive_like_every_downstream_reader(self) -> None:
-        # `_conductor_authors_runner` / `_pure_leaf_substep` /
-        # `_validate_toolchain_backend_supported` all compare the stripped value with no case
-        # folding. Lower-casing HERE would exempt `Infrastructure` at spec-input and then let
-        # all three treat it as a physics node — the removed leaf-authored-runner path, with
-        # no gate firing anywhere. It must fail closed instead.
-        for spelling in ("Infrastructure", "INFRASTRUCTURE", "infraStructure"):
-            self.assertIsNotNone(infra_dep_count_violation(spelling, 0), spelling)
-            self.assertIsNone(infra_dep_count_violation(spelling, 1), spelling)
-
-    def test_message_names_the_rule_the_count_and_the_canonical_doc(self) -> None:
-        msg = infra_dep_count_violation("component", 2)
-        self.assertIsNotNone(msg)
-        self.assertIn("exactly one", msg)
-        self.assertIn("infrastructure", msg)
-        self.assertIn("found 2", msg)
-        self.assertIn("docs/workflow/phases/phase_01_compile.md", msg)
-        # A remedy the author can act on, pointing the way the error actually goes: too
-        # many entries must be REMOVED, and telling the author to add one more would be
-        # actively wrong.
-        self.assertIn("Remove 1 of them", msg)
-        self.assertNotIn("Add the single", msg)
-        self.assertIn("advdiff1d_linear/deps.yaml", msg)
-        zero = infra_dep_count_violation("problem", 0)
-        self.assertIn("found 0", zero)
-        self.assertIn("Add the single `infrastructure_id` entry", zero)
-        self.assertIn("Remove 2 of them", infra_dep_count_violation("component", 3))
-
-    def test_unknown_or_non_string_kind_is_not_exempt(self) -> None:
-        # Only the literal `infrastructure` kind is exempt; anything else must declare one.
-        self.assertIsNotNone(infra_dep_count_violation(None, 0))
-        self.assertIsNotNone(infra_dep_count_violation("", 0))
-        self.assertIsNotNone(infra_dep_count_violation(123, 0))
-        self.assertIsNone(infra_dep_count_violation(None, 1))
 
 
 class LineWidthTest(unittest.TestCase):
